Create ExtractHighlightedWordText.py
Code designed to both extract highlighted text portions and recombine them. Often when extracting highlighted text from heavily edited or formatted docx files by using the <w:highlight> tags the extracted portions will be split into multiple parts. A highlighted sentence, such as, "this is a highlighted sentence," may be pulled from the doc as several elements, such as, "th","is","is","a highlighted sentence." This code is my attempt at pulling the sections and reconstructing them to their original form. I used difflib to compare the concatenated broken text to the original text and find similar sections. While difflib is an excellent library, this method is not perfect and misses out on some important highlighted portions as a result.