Create Julia_Set_Movie.py
The Mandelbrot set is the set of complex numbers, c, for which the function f(z)=z^2+c does not diverge when iterated from z=0. Plotting the results generates a fractal. This set is closely related the to the Julia set. This code plots a multi-julia set through a similar iterative process to that described. It repeats this process with varying inputs to create multiple images of multi-julia sets and compiles these images into a .gif to create a short animation cycling through multi-julia sets.